[homework-7] First attempt

+++ library_data_jpa/src/main/java/ru/otus/homework7/domain/Comment.java
@@ -23,11 +23,11 @@ public class Comment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private int id;
+    private long id;
 
     @Column(name = "comment")
     private String comment;
 
     @Column(name = "book_id")
-    private int bookId;
+    private long bookId;
 }
