+++ library_data_jpa/src/main/java/ru/otus/homework7/service/StyleService.java
@@ -3,6 +3,6 @@ package ru.otus.homework7.service;
 import ru.otus.homework7.domain.Style;
 
 public interface StyleService {
-    Style getStyleById(int styleId);
+    Style getStyleById(long styleId);
     Style getStyleByName(String styleName);
 }
